Move everything to a module for Paranoid, since in the future I might want multiple variants
