+++ battlesnake-minimax/src/paranoid/minimax_return.rs
@@ -2,7 +2,7 @@ use battlesnake_game_types::types::{Move, SnakeIDGettableGame};
 use std::fmt::Debug;
 use text_trees::StringTreeNode;
 
-use crate::WrappedScore;
+use super::WrappedScore;
 
 #[derive(Debug, Clone)]
 /// This is returned from an iteration of the minimax algorithm
