+++ battlesnake-minimax/src/paranoid/score.rs
